+++ src/main/java/net/yxiao233/ifeu/common/registry/ModRarity.java
@@ -1,0 +1,8 @@
+package net.yxiao233.ifeu.common.registry;
+
+import net.minecraft.ChatFormatting;
+import net.minecraft.world.item.Rarity;
+
+public class ModRarity {
+    public static final Rarity LEGENDARY = Rarity.create("legendary",ChatFormatting.GOLD);
+}